refactor(kolme): extract in-memory runtime client module file (#1938)
